requirements.txt
# ════════════════════════════════════════════════════════════════
# CONCEPTION — THE BUILDER FOUNDRY
# requirements.txt — Phase 1 (all phases backward-compatible)
# ════════════════════════════════════════════════════════════════

# ── Core web framework ──
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0

# ── HTTP client (service-to-service + external APIs) ──
httpx>=0.27.0

# ── Database ──
psycopg2-binary>=2.9.9

# ── Redis + Celery (task queue for forge/scan workers) ──
redis>=5.0.0
celery>=5.3.0

# ── Auth ──
PyJWT>=2.8.0
python-dotenv>=1.0.0

# ── Payments ──
stripe>=8.0.0

# ── AI backbone ──
anthropic>=0.25.0
google-generativeai>=0.7.0

# ── Streamlit (dashboard UI) ──
streamlit>=1.35.0

# ── Utilities ──
python-dateutil>=2.9.0
